+++ acidcool_detect_audio_language/lib/ffmpeg/probe.py
@@ -4,7 +4,7 @@
 """
     plugins.probe.py
 
-    Written by:               Josh.5 <jsunnex@gmail.com>
+    Patched by:               AciDCooL (Forked from Josh.5)
     Date:                     12 Aug 2021, (9:20 AM)
 
     Copyright: